Move integrator function into controller class
